merge to buyer

+++ app/Http/Controllers/seller/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\seller;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Product;
+use App\Models\User;
 use App\Http\Requests\seller\ProductRequest;
 class ProductController extends Controller
 {
@@ -13,11 +14,12 @@ class ProductController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $product=Product::all();
 
-        return view('seller.sellerProducts',compact('product'));
+        $user=User::find($request->session()->get('id'));
+        $product=Product::where('seller_id',$user->id)->get();
+        return view('seller.sellerProducts',compact('product','user'));
     }
 
     /**
@@ -25,9 +27,10 @@ class ProductController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
+    public function create(Request $request)
     {
-        return view('seller.createproduct');
+        $user=User::find($request->session()->get('id'));
+        return view('seller.createproduct',compact('user'));
     }
 
     /**
@@ -39,32 +42,43 @@ class ProductController extends Controller
     public function store(ProductRequest $request)
     {
         $product = new Product;
+        $user=User::find($request->session()->get('id'));
+        if($user->points >=10 || $user->prime_status=="prime"){
+            if($user->prime_status!="prime")
+            $user->points=$user->points-10;
+            $user->update();
+            if($request->hasFile('product_picture')){
+                $extension = $request->product_picture->getClientOriginalExtension();
+                $newName = date('U').'.'.$extension;
+                $folderPath = "seller/image/product/";
+                $product->product_picture = $folderPath.$newName;
+                $request->product_picture->move($folderPath, $newName);
+            }
 
-        if($request->hasFile('product_picture')){
-            $extension = $request->product_picture->getClientOriginalExtension();
-            $newName = date('U').'.'.$extension;
-            $folderPath = "seller/image/product/";
-            $product->product_picture = $folderPath.$newName;
-            $request->product_picture->move($folderPath, $newName);
+            $product->name= $request->input('name');
+            $product->price= $request->input('price');
+            $product->description= $request->input('description');
+            $product->number_of_info= $request->input('number_of_info');
+            $product->Pyament_recive_no= $request->input('Pyament_recive_no');
+            // $product->delete_status= $request->input('delete_status');
+
+            $product->from_currency= $request->input('from_currency');
+            $product->To_currency= $request->input('To_currency');
+            $product->seller_id=$request->session()->get('id');
+            if($product->save()){
+                $request->session()->flash('msg',"Product Added Successfully!");
+            }
+            else
+            {
+                $request->session()->flash('msg'," Failed To Add Product!");
+            }
+
+        }
+        else{
+            $request->session()->flash('msg'," you do not have enough points, you can upgrate to prime seller!");
         }
 
-        $product->name= $request->input('name');
-        $product->price= $request->input('price');
-        $product->description= $request->input('description');
-        $product->number_of_info= $request->input('number_of_info');
-        $product->Pyament_recive_no= $request->input('Pyament_recive_no');
-        // $product->delete_status= $request->input('delete_status');
 
-        $product->from_currency= $request->input('from_currency');
-        $product->To_currency= $request->input('To_currency');
-        $product->seller_id=1;
-        if($product->save()){
-            $request->session()->flash('msg',"Product Added Successfully!");
-        }
-        else
-        {
-            $request->session()->flash('msg'," Failed To Add Product!");
-        }
 
          return redirect()->Back();
     }
@@ -75,12 +89,13 @@ class ProductController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show(Product $product)
+    public function show(Product $product,Request $request)
     {
+        $user=User::find($request->session()->get('id'));
         $payment_methods = array('none',"Bkash", "Nagod", "roket","Mkash","Ukash","Gkash");
         $counter=0;
         $counter2=0;
-        return View('seller.showproduct',compact('product','payment_methods','counter','counter2'));
+        return View('seller.showproduct',compact('product','payment_methods','counter','counter2','user'));
 
     }
 
@@ -90,13 +105,13 @@ class ProductController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit(Product $product)
-
+    public function edit(Product $product,Request $request)
     {
+        $user=User::find($request->session()->get('id'));
         $payment_methods = array('none',"Bkash", "Nagod", "roket","Mkash","Ukash","Gkash");
         $counter=0;
         $counter2=0;
-        return View('seller.editproduct',compact('product','payment_methods','counter','counter2'));
+        return View('seller.editproduct',compact('product','payment_methods','counter','counter2','user'));
     }
 
     /**
@@ -108,31 +123,36 @@ class ProductController extends Controller
      */
     public function update(ProductRequest $request, $id)
     {
+         $user=User::find($request->session()->get('id'));
+            if($user->points >=5 && $user->prime_status=="prime"){
+                if($user->prime_status!="prime"){
+                    $user->points=$user->points-5;
+                }
+                $product=Product::find($id);
+                if($request->hasFile('product_picture')){
+                if($product->product_picture)unlink($product->product_picture);
+                $extension = $request->product_picture->getClientOriginalExtension();
+                $newName = date('U').'.'.$extension;
+                $folderPath = "seller/image/product/";
+                $product->product_picture = $folderPath.$newName;
+                $request->product_picture->move($folderPath, $newName);
+                }
+                $product->name= $request->input('name');
+                $product->price= $request->input('price');
+                $product->description= $request->input('description');
+                $product->number_of_info= $request->input('number_of_info');
+                $product->Pyament_recive_no= $request->input('Pyament_recive_no');
+                $product->from_currency= $request->input('from_currency');
+                $product->To_currency= $request->input('To_currency');
+                $product->number_of_info=$request->input('number_of_info');
+                $product->update();
+                $request->session()->flash('msg','Product is Updated!');
+            }
+            else{
+                $request->session()->flash('msg'," you do not have enough points");
+             }
 
-        $product=Product::find($id);
-        if($request->hasFile('product_picture')){
-        if($product->product_picture)unlink($product->product_picture);
-        $extension = $request->product_picture->getClientOriginalExtension();
-        $newName = date('U').'.'.$extension;
-        $folderPath = "seller/image/product/";
-        $product->product_picture = $folderPath.$newName;
-        $request->product_picture->move($folderPath, $newName);
-        }
 
-
-        $product->name= $request->input('name');
-        $product->price= $request->input('price');
-        // $product->product_picture= $request->input('product_picture');
-        $product->description= $request->input('description');
-        $product->number_of_info= $request->input('number_of_info');
-        $product->Pyament_recive_no= $request->input('Pyament_recive_no');
-        // $product->delete_status= $request->input('delete_status');
-        $product->from_currency= $request->input('from_currency');
-        $product->To_currency= $request->input('To_currency');
-        $product->number_of_info=$request->input('number_of_info');
-        $product->update();
-        $request->session()->flash('msg','Product is Updated!');
-        // return redirect()->route('seller.product.show',$product);
         return redirect()->back();
     }
 
@@ -142,7 +162,7 @@ class ProductController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function active(ProductRequest $request, $id )
+    public function active(Request $request, $id )
     {   $product=Product::find($id);
         $product->delete_status= 'active';
         $product->update();
@@ -150,7 +170,7 @@ class ProductController extends Controller
         return redirect()->back();
     }
 
-    public function deactive(ProductRequest $request, $id )
+    public function deactive(Request $request, $id )
     {   $product=Product::find($id);
         $product->delete_status= 'deactive';
         $product->update();
@@ -161,7 +181,7 @@ class ProductController extends Controller
 
 
 
-    public function destroy(Product $product, ProductRequest $request)
+    public function destroy(Product $product,Request $request)
     {
         $product->delete_status= 'deleted';
         $product->update();
